Initial weta waf+cmake build setup, with standard pipelines
There are a number of tests that are failing, mainly due to assumptions in cmake in how the tests are being run. These are disabled + documented within the wscript.

In addition, the opensource CI pipeline are removed, and replaced with our standard pipeline, with one addition. We triger both the provider and wak-cmake in order to validate the new version works correctly.

Signed-off-by: Niall Lenihan <nlenihan@wetafx.co.nz>