Expose the module file path that was used to attempt finding a package.
At the moment there is no way of knowing if loading from a module was attempted and failed, or if no module was found.

We have a provider that will fall back to a different method only if no config and no module were located. Therefore, it is importatnt to know if a module file was found or not.

This mirrors the `<PackageName>_CONFIG` variable, so it is consistent with how cmake is operating for config mode.

Signed-off-by: Niall Lenihan <nlenihan@wetafx.co.nz>