Enable building the cmake gui.
The cmake gui is very usful, and distributing it along side our standard cmake build can offer our developers another introspection tool.

While UI is VP specific (as it uses Qt), this is a standalone app and is not required to be used from within VP specific workflows. However, The path of least resistance is to update the ABI based build to VP, and use only VP23.

However, this introduces a small wrinkle
* We should not require Qt when bringing in cmake, as it is only required for the UI. Therefore, exposing this as an additional pak is preferred, as this allows specific requirements to be set.
* All of the built executables are placed in a single directory, so the existing cmake pak will expose the Qt dependent app.

To get around this issue, I have moved the install location of the gui to make it unique. This means the existing cmake pak can expose the existing functionality without requiring Qt/VP at the pak level, while the cmake-gui pak can expose the Qt dependant UI.

Signed-off-by: Niall Lenihan <nlenihan@wetafx.co.nz>